backend - xmeme/requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
psycopg2==2.8.6
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
